+++ Application/Model/ManagerHandler.php
@@ -8,35 +8,6 @@ use OxidEsales\Eshop\Core\ViewConfig;
 class ManagerHandler
 {
     /**
-     * Gets current chosen Manager
-     *
-     * @return string
-     */
-    public function getCurrManager() :string
-    {
-        /** @var ManagerTypes $oManagerTypes */
-        $oManagerTypes = oxNew(ManagerTypes::class);
-
-        /** @var ViewConfig $oViewConfig */
-        $oViewConfig = oxNew(ViewConfig::class);
-
-        $aManagerList =  $oManagerTypes->getManagerList();
-
-        if ($this->getModuleSettingExplicitManagerSelectValue()){
-            return $this->getExplicitManager();
-        }
-
-        foreach ($aManagerList as $shopModuleId => $publicCMPName){
-           if ($oViewConfig->isModuleActive($shopModuleId)){
-               $this->d3SaveShopConfVar($shopModuleId);
-               return $shopModuleId;
-           }
-        }
-
-        return "";
-    }
-
-    /**
      * @param string $sParam
      * @return void
      */
@@ -53,26 +24,8 @@ class ManagerHandler
     /**
      * @return string
      */
-    public function getModuleSettingExplicitManagerSelectValue() :string
+    public function getActManager() :string
     {
-        return Registry::get(ViewConfig::class)->d3GetModuleConfigParam('_HAS_STD_MANAGER')?:"";
-    }
-
-    /**
-     * @return string
-     */
-    public function getExplicitManager() :string
-    {
-        $sPotentialManagerName = $this->getModuleSettingExplicitManagerSelectValue();
-
-        /** @var ManagerTypes $oManagerTypes */
-        $oManagerTypes = oxNew(ManagerTypes::class);
-        $sCMPName = $oManagerTypes->isManagerInList($sPotentialManagerName)
-            ? $sPotentialManagerName
-            : "NONE";
-
-        $this->d3SaveShopConfVar($sCMPName);
-
-        return $sCMPName;
+		return Registry::get(ViewConfig::class)->d3GetModuleConfigParam('_HAS_STD_MANAGER')?:"";
     }
 }